backend-web-service/requirements.txt
sanic==25.3.0
sanic-cors==2.2.0
openai==2.5.0
python-dotenv==1.1.1
spotipy==2.25.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.4